+++ gitnexus/src/config/supported-languages.ts
@@ -9,7 +9,7 @@
  *    ----------------------------------|------------------------------------------|---------------------------
  *    tree-sitter-queries.ts            | Query string + LANGUAGE_QUERIES entry    | (required)
  *    export-detection.ts               | ExportChecker function + table entry     | (required)
- *    import-resolution.ts              | Resolver in buildImportResolvers         | resolveStandard(...)
+ *    import-resolution.ts              | Resolver in importResolvers              | resolveStandard(...)
  *    import-resolution.ts              | namedBindingExtractors entry             | undefined
  *    call-routing.ts                   | callRouters entry                        | noRouting
  *    entry-point-scoring.ts            | ENTRY_POINT_PATTERNS entry               | []
