+++ gitnexus/src/core/ingestion/type-extractors/index.ts
@@ -47,6 +47,5 @@ export {
   extractSimpleTypeName,
   extractGenericTypeArgs,
   extractVarName,
-  findChildByType,
   extractRubyConstructorAssignment
 } from './shared.js';
