+++ gitnexus/test/fixtures/lang-resolution/csharp-no-csproj/Models/User.cs
@@ -1,0 +1,8 @@
+namespace Models
+{
+    public class User
+    {
+        public void Save() { }
+        public string GetName() { return "alice"; }
+    }
+}
