+++ gitnexus/test/fixtures/lang-resolution/csharp-no-csproj/Services/UserService.cs
@@ -1,0 +1,13 @@
+using Models;
+
+namespace Services
+{
+    public class UserService
+    {
+        public void ProcessUser()
+        {
+            var user = new User();
+            user.Save();
+        }
+    }
+}
