+++ gitnexus/test/fixtures/lang-resolution/go-cmd-helper/cmd/server/internal/config/config.go
@@ -1,0 +1,5 @@
+package config
+
+func Load() string {
+	return "loaded"
+}
