+++ gitnexus/test/fixtures/lang-resolution/go-cmd-helper/cmd/server/main.go
@@ -1,0 +1,7 @@
+package main
+
+import "myapp/cmd/server/internal/config"
+
+func main() {
+	config.Load()
+}
