+++ gitnexus/test/fixtures/lang-resolution/php-use-function-const/app/Config/constants.php
@@ -1,0 +1,5 @@
+<?php
+
+namespace App\Config;
+
+const MAX_RETRIES = 3;
