+++ gitnexus/test/fixtures/lang-resolution/php-use-function-const/app/Models/User.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace App\Models;
+
+class User {
+    public function save(): void {}
+    public function getName(): string {
+        return "alice";
+    }
+}
