+++ gitnexus/test/fixtures/lang-resolution/php-use-function-const/app/Services/Calculator.php
@@ -1,0 +1,17 @@
+<?php
+
+namespace App\Services;
+
+use App\Models\User;
+use function App\Utils\formatName;
+use const App\Config\MAX_RETRIES;
+
+class Calculator {
+    public function process(): void {
+        $user = new User();
+        $user->save();
+
+        $name = formatName("test");
+        echo MAX_RETRIES;
+    }
+}
