+++ gitnexus/test/fixtures/lang-resolution/php-use-function-const/app/Utils/helpers.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace App\Utils;
+
+function formatName(string $name): string {
+    return strtoupper($name);
+}
