+++ gitnexus/test/fixtures/lang-resolution/rust-scoped-multi-file/src/models/mod.rs
@@ -1,0 +1,5 @@
+mod user;
+mod repo;
+
+pub use user::User;
+pub use repo::Repo;
