+++ gitnexus/test/fixtures/lang-resolution/rust-scoped-multi-file/src/models/repo.rs
@@ -1,0 +1,13 @@
+pub struct Repo {
+    name: String,
+}
+
+impl Repo {
+    pub fn new(name: &str) -> Self {
+        Repo { name: name.to_string() }
+    }
+
+    pub fn clone_repo(&self) {
+        println!("Cloning repo {}", self.name);
+    }
+}
