+++ gitnexus/test/fixtures/lang-resolution/rust-scoped-multi-file/src/models/user.rs
@@ -1,0 +1,13 @@
+pub struct User {
+    name: String,
+}
+
+impl User {
+    pub fn new(name: &str) -> Self {
+        User { name: name.to_string() }
+    }
+
+    pub fn save(&self) {
+        println!("Saving user {}", self.name);
+    }
+}
